Fix StyleCop
