+++ Microsoft.Toolkit.Uwp.UI.Controls.Graph/AadLogin/AadLogin.Properties.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using Microsoft.Toolkit.Services.MicrosoftGraph;
-using System;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Media.Imaging;
